Updated for Optimal Functionality
Updated my models to include prisons and convicts, as well as connecting convicts to prisons in the model.

Created my Prison Lookup HTML template, fixed the defendant and judge lookups to request little information so as to be efficient to a user.

Created case lookup, defendant lookup, and judge lookup views.

Tweaked forms and admin files for optimal functionality.